Add info/extract.py initial shell
Markdown and callbacks all seem to be working pretty well.

It seems like it's going to be clean enough just to write the
Markdown in the source file. Multi-line string is easy enough.
Big question will be whether the Markdown reflow will work
smoothly with the Markdown buried in strings like that....